Adding 3rd field i.e. Payment Type - Cash / Online
Both in UI in Actionsheet Controller
and in the modal we change the UI and make it alive.
Now in database we add 3rd columns i.e. credit_type for saving the form input.

This commit ensure that our 3rd field i.e. Payment type successfully implement into our application from UI to database that's it.