Popup-Input widget from scratch ( copy the iphone theme )
In this way, we use the modal and take the input from it.
I try to make it as a iphone input box and it's looking fine.
Next we create the data inside our database if the input is valid.
Actually it's the runs the 'CREATE' query in the database.