UPDATING through same component (PopupInput) used by INSERTING.
PopupInput, same component is used for both updating and inserting.
All we need to do is just one condition to check weather the ID is there or not.
If ID there this means it's should updating in the database.
Otherwise it's inserting.
--------------------------------------------------------------
Before inserting we make sure the ID is zero to make if statemetn false.