Use ActionSheet for popup message
Just adding a condition outside of flatlist means if the data is not provided for iteration then entire list space disappear.
This makes our Sheet more meaninful for showing short messages.
By adding a condition we just remove all the unncesscary padding around the flatlist item if data is not provided.