Others - Default Source, & user can't delete it (hide it in edit page)
The "Others" is the default source creating by the application for us.
In this commit,we ensure user can't delete this item because it's the application default value.
So we hides it in the edit page.
But it shown in Drop-down menu that's the beauty of coding.
We hides alot of stuff behind the curtain.