HomeScreen - Cash, Online & Total Balance concept complete.
Now total balance with the separation of how much of it is online/cash is shown.
Adding 2 fields in the Pocket Table i.e. cashBal and onlineBal and insert the default value at the starting of the application.
-----------------------------------------------------------------------
readingPocket() change code same as what we do in INSERTION of "Others" in source table
i.e. IGNORE or INSERT ( the game )
-------------------------------------------------------------------
Testing application working perfectly fine.