(Important) Transaction Screen's Segment Button
These button forces the component to re-render because we re-run the query after some modification.
All , Credit , Debit | these are segment buttons.
---------------------------------------------------------------------
Transaction don't load on clicking the tabbar because it's conflict with this feature therefore we remove that section of code in this commit.
On changing the data we will pull to refresh the data from the database 
(PULL TO REFRESH )
---------------------------------------------------------------------
Double check program runs perfectly fine.
May be in future we will add the loading component.