Refreshing the Transaction list feature is successfully added.
It's working fabulusly fine because i just debug it right now.
Transaction list working good, on first time the app opens - it's shows all data, on changing the navigation    we successfully show that tab bar data "All" , 'Credit".
----------------------------------------------------------------
Game - The game is on refreshing the page the new data coming from the database and the segment button don't change.
Means child don't refresh on changing the state in parent component.