Adding Delete Section to Transaction ( UI is not ready )
After this commit we can delete the transaction which we don't want and it will also handle the pocket related to it.
-------------------------------------------------------
Point to be noted the UI like how we access the delete button is not ready, we will add SLIDER later to each flatlist item.
------------------------------------------------------------

It's working fine :)